requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.3.0.23
bleach==3.2.1
celery==3.1.23
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1
django-bootstrap3==12.1.0
django-crispy-forms==1.9.0
django-filter==21.1
django-model-utils==3.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.3
entrypoints==0.3
Faker==4.1.0
filelock==3.0.12
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.2
gTTS==2.2.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
inflection==0.5.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.14
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.0
Js2Py==0.70
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kombu==3.0.35
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.3
misaka==2.1.1
mistune==0.8.4
more-itertools==8.6.0
MouseInfo==0.1.3
mpmath==1.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.0
nest-asyncio==1.4.3
nose==1.3.7
notebook==6.2.0
numpy==1.18.1
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==7.0.0
pipwin==0.5.1
playsound==1.2.2
pocketsphinx==0.1.15
prometheus-client==0.9.0
prompt-toolkit==3.0.10
psycopg2==2.8.4
py4j==0.10.9
PyAudio @ file:///C:/Users/ftkyu/pipwin/PyAudio-0.2.11-cp38-cp38-win32.whl
PyAutoGUI==0.9.52
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.9
Pygments==2.7.4
pyjsparser==2.7.1
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.2
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
pyserial==3.5
pySmartDL==1.3.4
pyspark==3.1.2
python-dateutil==2.8.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.5
pywhatkit==3.9
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
Quandl==3.6.0
redis==2.10.5
requests==2.24.0
scikit-learn==0.24.0
scipy==1.6.0
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
serial==0.0.97
six==1.15.0
sklearn==0.0
soupsieve==2.2
SpeechRecognition==3.8.1
SQLAlchemy==1.4.26
sqlparse==0.3.0
sympy==1.7.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
twilio==6.45.0
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.29
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
widgetsnbextension==3.5.1
wikipedia==1.4.0
WTForms==2.3.3
xlrd==2.0.1
